Rewrite calculation of 2D morphology for images
Rewrite calculation of 2D morphology for drawing images to use only
code used for finding indices for CSD and potential transformations
from loops to 3D morphology, to assure that morphology and CSD (or
potential) will match perfectly. Provide extent in ordinary form
and not switched x and y (for easier use with imshow).