New atribute in KCSD class
I would like to include new auxiliary atribute of the KCSD class "own_est" which is an array of shape (number of dimensions 2 or 3, number of csd points coordinates) .
It allows to reconstruct CSD only in given set of points (i.e. body of neuron) and work only when the array is not empty (default value is np.array([])).