Use loop right ends as estimation points for sKCSD
There was an inconsitency in choosing points for sKCSD estimation. The
point of origin was also taken, which made the x dimension k_interp_pot
 different from the number of loops. This has been fixed by adding a new
class variable, loop_pos, which is used only for sKCSD estimation.
Variables est_poz and est_xyz are used for finding Cartesian coordinates
of points on the morphology loop.