Seperation of src and est locations
There are two corrections in the code. In method suggest Lambda there was a mistake in taking maximal value of lambda (should be log10 from that value).
Second in classes oKCSD2D and 3D from now on locations of sources and estimations of potential/csd are  seperated and are called own_src and own_est respectively.  if only own_src are used own_est will have default positions the same as are in own_src array.