Update estm_dists tests
estm_dists are of the len(sKCSDcell.source_pos) x len(sKCSD.est_pos) - 1
size. sKCSD.est_pos contains all the loop positions starting with the
origin, which results in two points for the first loop and one point for
all the others. Consequently the origin is not included in estm_dists.